engine: q-mix draw data (slewedQmix + QmixDraw + redraw/recompute), replace STRAND_QMIX placeholders, wire Dice/Mix/Slew Q + QMIX_LEVEL
Task 4b. Mirror the melody stream 1:1 for q-mix on its own STREAM_SOURCE_SELECT Philox. Adds slewedQmix[16]+slewedPolyQmix[15][16]+qmixSource/polyQmixSource, QmixDraw + patternQmixAt/rawDrawQmixPatternAt, recomputeEffectiveQmix + redrawQmix, pending-dice dispatch + mix/slew recompute, setPendingQmix*/isQmixSeedPending. Replaces the STRAND_QMIX/PL_QMIX 0.5f placeholders in remapSlewedByPins/pickMono with real pins that promote into qmixRandom/polyRandom(PL_QMIX) exactly like melody. latchMix extended with qmix mix/slew (all callers + unit test updated); PatternInput gains qmixSlew/qmixMix/qmixLevel. MonsoonModeController feeds them in; Monsoon::diceQmix()/Last-Dice Q call setPendingQmixRoll/LastRoll. QMIX_LEVEL: getQmixLevel() + SequencerEngine thresholds the mono q-mix draw (r_qmix<level -> hit) mirroring accent. Also casts the Bug-B editor-enum-vs-strand static_asserts to int to silence -Wenum-compare. Engine verified: Dice Q arms + redraws q-mix at the phrase boundary, decorrelated from melody; compile clean. (QMIX on-screen bars + spread + a Dice Q pending light land in 4c.)

+++ src/ui/SandsVisualEditorV4.hpp
@@ -59,10 +59,13 @@ struct SandsVisualEditorV4 : rack::TransparentWidget {
   static_assert(MELODY == 0 && OCTAVE == 1 && QMIX == 2 && REST == 3
                 && ACCENT == 4 && VARIATION == 5 && LEGATO == 6,
                 "SandsVisualEditorV4::Lane must equal the QMIX editor order (dsp/LaneMapping.hpp)");
-  static_assert(MELODY == dotModular::STRAND_MELODY && OCTAVE == dotModular::STRAND_OCTAVE
-                && QMIX == dotModular::STRAND_QMIX && REST == dotModular::STRAND_RHYTHM
-                && ACCENT == dotModular::STRAND_ACCENT && VARIATION == dotModular::STRAND_VARIATION
-                && LEGATO == dotModular::STRAND_LEGATO,
+  // (int) casts: Lane and dotModular::EngineStrand are distinct enums; comparing them directly
+  // trips -Wenum-compare. We deliberately assert their integer values coincide (editor lane ==
+  // strand, MONO_LANE_TO_STRAND identity), so compare as ints.
+  static_assert((int)MELODY == (int)dotModular::STRAND_MELODY && (int)OCTAVE == (int)dotModular::STRAND_OCTAVE
+                && (int)QMIX == (int)dotModular::STRAND_QMIX && (int)REST == (int)dotModular::STRAND_RHYTHM
+                && (int)ACCENT == (int)dotModular::STRAND_ACCENT && (int)VARIATION == (int)dotModular::STRAND_VARIATION
+                && (int)LEGATO == (int)dotModular::STRAND_LEGATO,
                 "editor lane == engine strand (MONO_LANE_TO_STRAND is identity); keep both in step");
   
   struct Colors {
